requirements.txt
fastapi
uvicorn[standard]
pydantic
pymongo
python-dotenv
requests
serpapi
google-search-results
